revert to manual priority setting; debug

+++ eth_priority_tests/sink.c
@@ -245,7 +245,7 @@ void thread_recv_source_data()
                 print_hex((char*) &frame, 40); printf("\n");
                 print_hex(frame.payload.data+4, 40); printf("\n");
 
-                printf("alignment diff from frame start to frame data: %d", &(frame.payload) - &frame);
+                printf("alignment diff from frame start to frame data: %d", ((int)&(frame.payload)) - ((int)&frame));
 
                 memcpy(&time_from_source, &(frame.payload.ss_payload.tx_time), sizeof(struct timespec));
